JS code cleanup in No data and Tracking code page, #PG-2799
